+++ CHANGELOG.md
@@ -8,8 +8,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-### Added
-
+- added `delete` method to `FREDJob` class.
 - added a method to FREDRun, get_table_variable, that loads table varibale out for a FRED run.
 
 ### Fixed